+++ lib/core/models/video/video_render_data_model.dart
@@ -456,6 +456,7 @@ class VideoRenderData {
                 'height': layer.size?.height,
                 'rotation': layer.rotation,
                 'loop': layer.loop,
+                'animationOffsetUs': layer.animationOffset.inMicroseconds,
                 'animations': layer.animations.map((a) => a.toMap()).toList(),
               },
             ),